+++ src/chameleon_log/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.9.0'
+__version__ = '2.0.0'
 
 from .detectors import get_log_handler, is_connected_journald
 from .rich import RichHandler